Refactor CLI into seperate package
